more refactoring of diary page;
